+++ src/index.php
@@ -18,13 +18,11 @@ if (strncasecmp($url,$admin_prefix, strlen($admin_prefix)) === 0) {
     $url = str_replace($admin_prefix, '', preg_replace('/\?.+/', '', $url));
     $app = new Box_AppAdmin();
     $app->setUrl($url);
-    $di['translate']('admin');
 } else {
     $app = new Box_AppClient();
     $app->setUrl($url);
-    $di['translate']();
 }
-
+$di['translate']();
 $app->setDi($di);
 print $app->run();
 exit(); // disable auto_append_file directive